refactored the device function
